requirements.txt
transformers==4.42.4
pytorch_lightning==2.3.3
numpy==1.23.1
pandas==1.4.4
scikit-learn==1.3.2
python-dotenv==1.0.1
Pebble==5.0.7
timeout-decorator==0.5.0
einops==0.8.0
antlr4-python3-runtime==4.13.1
datasets==2.20.0
accelerate==0.32.1
flash_attn==2.6.1
trl==0.9.6